Fixed most errors, only some undefined references left
